requirements.txt
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
requests==2.28.2
opentelemetry-instrumentation-requests==0.36b0
